requirements.txt
async-timeout==4.0.2
Deprecated==1.2.13
mrz==0.6.2
packaging==21.3
pyaes==1.6.1
pyparsing==3.0.8
Pyrogram==1.4.16
PySocks==1.7.1
python-decouple==3.6
TgCrypto==1.2.3
wrapt==1.14.0
